refactor: change structure to not intercept with pages routing
