+++ src/app/AboutPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from '../../pages/AboutPage/ui/AboutPage';
