+++ src/app/AboutPage/ui/AboutPage.tsx
