+++ src/app/DetailedCardPage/index.ts
