+++ src/app/DetailedCardPage/ui/DetailedCardPage.module.css
