+++ src/app/DetailedCardPage/ui/DetailedCardPage.tsx
