+++ src/app/MainPage/ui/MainPage.tsx
