+++ src/app/[locale]/about/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next';
-import AboutPage from '../../../pages/AboutPage';
+import AboutPage from '../../AboutPage';
 
 export const metadata: Metadata = {
   title: 'PokéDexplorer. About',
